requirements.txt
fastapi~=0.98.0
pydantic~=1.10.9
starlette==0.28.0
python-jose
passlib~=1.7.4
bcrypt~=4.0.1
python-dotenv~=1.0.0
uvicorn~=0.22.0
python-multipart